extracted logic to a separate custom hook, reuse the useproductsbycategory hook
