+++ .eslintrc.json
@@ -7,7 +7,10 @@
     "plugin:react-hooks/recommended",
     "prettier"
   ],
-  "rules": { "react/prop-types": 0, "react/react-in-jsx-scope": 0 },
+  "rules": {
+    "react/prop-types": 0,
+    "react/react-in-jsx-scope": 0
+  },
   "plugins": ["react", "import", "jsx-a11y"],
   "parserOptions": {
     "ecmaVersion": 2022,
